Add heuristic to determine location of PyFrameObject
DockerHub CPython images are now built with Profile Guided Optimizations
enabled which breaks discovering of PyFrameObject's. Unlike with
partially optimized out frames, this completely breaks `py-bt` as it
now gives an empty result.

Add heuristic that allows to determine the location of PyFrameObject
in PyEval_EvalFrameDefault frames: it looks like it always ends up
in one of the general purpose CPU registers, so we can just try each
one until we find a proper PyObject with type "frame". Unfortunately,
this is a bit fragile as sometimes there can be two PyFrameObject's -
one for the current frame and one for the child frame. In this case
we can use the `f_back` pointer in PyFrameObject to determine the
one we need.

I would normally not add such heuristic, but w/o this hack cpython-lldb
is currently unusable for DockerHub CPython builds.