Run tests on CPython 3.8 as well
CPython 3.8 has been released, so let's make sure it's tested too.

Either due to LTO optimizations enabled, or due to the new fast
calling protocol (PEP 590) implemented in CPython 3.8, we can no
longer get a proper representation of builtin_id() frame in LLDB,
that breaks the pretty-printer tests. To fix the tests, we can use
the fact that according to the rules of AMD64 calling convention
the argument of builtin_id() is passed in the CPU register RSI.