Implement a command for listing of the source code
With a caveat that at the moment it's only possible to list the
source code of the Python module being executed in the most recent
Python frame. This will be fixed when we add support for walking up
and down the Python call stack.

Small fixes for Python 3 compatibility (i.e. when LLDB is linked
against Python 3 libpython).