Run tests in parallel
Tests are rather slow now. We can look into reducing the number of
times LLDB is started, but the simplest optimization can be running
the tests in parallel by the means of pytest-xdist.

Two minor complications that had to be resolved:

1) a temporary directory was created for each test, but chdir() was
never executed. Fix that.

2) strip_symbols temporary modifies the Python binary in-place, which
obviously does not play nicely with other tests when they are run
in parallel. Work around that by running the single test we use this
fixture in separately