testing: reduce the test run time by not restarting LLDB
Instead of starting LLDB for each test case, we can start it once in
the background and reuse it across all the tests run by a given pytest-xdist
process. This requires each test to clean up the breakpoints and restart
the process under debug, but overall, it's not too hard. Communication
with the LLDB instance happens via pipes attached to its stdin/stdout/stderr
and is managed by pexpect.

This, combined with the increased concurrency level, allowed to reduce
the time it takes to complete a single test run from ~45s to ~24s (on
my machine). Unfortunately, that does not translate to a speed up on CI,
as the Github Actions runner seem to only have a single CPU core.