Use the C-extension for testing pretty printers
The tests started to fail on Python 3.10 + LLDB 11 recently. The issue
is with that particular CPython build from the corresponding official
Docker image: the interpreter is built with all optimizations enabled,
and builtin_id() is inlined, which means our breakpoint is no longer
hit.

Previously, we had to resort to relying on the AMD64 calling convention
for finding the variable we try to pretty-print in tests. That worked,
but was unnecessarily annoying.

To avoid all that trouble, create a helper function in the C-extension
that we already use for testing. Here we have full control over the
compiler options used and we can disable optimizations to get simpler
and more reliable tests.